backend/requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.0.1
Flask-CORS==4.0.0
python-docx==0.8.11
mysql-connector-python==8.1.0
Werkzeug==2.3.7